fix circular import errors. don't propagate logging levels within model, since that means that when merging configs from multiple sources we end up taking default values from higher priority sources when unset
