+++ miniscope_io/models/config.py
@@ -61,17 +61,6 @@ class LogConfig(MiniscopeIOModel):
             value = value.upper()
         return value
 
-    @model_validator(mode="after")
-    def inherit_base_level(self) -> "LogConfig":
-        """
-        If loglevels for specific output streams are unset, set from base :attr:`.level`
-        """
-        levels = ("level_file", "level_stdout")
-        for level_name in levels:
-            if getattr(self, level_name) is None:
-                setattr(self, level_name, self.level)
-        return self
-
 
 class Config(BaseSettings, YAMLMixin):
     """
